+++ js/lib/flynn/flynnParticles.js
@@ -27,22 +27,22 @@ Flynn.Particle = Class.extend({
         this.life = this.PARTICLE_LIFE + (Math.random()-0.5) * this.PARTICLE_LIFE_VARIATION;
     },
 
-    update: function(paceFactor) {
+    update: function(elapsedTicks) {
         var isAlive = true;
         // Decay and die
-        this.life -= paceFactor;
+        this.life -= elapsedTicks;
         if(this.life <= 0){
             // Kill particle
             isAlive = false;
         }
         else{
             // Add impulse
-            this.position.add(this.velocity.clone().multiplyScalar(paceFactor));
+            this.position.add(this.velocity.clone().multiplyScalar(elapsedTicks));
             // Decay impulse
-            this.velocity.multiplyScalar(Math.pow(this.PARTICLE_FRICTION, paceFactor));
+            this.velocity.multiplyScalar(Math.pow(this.PARTICLE_FRICTION, elapsedTicks));
             if(this.length !== null){
                 // Rotate line
-                this.angle += this.angular_velocity * paceFactor;
+                this.angle += this.angular_velocity * elapsedTicks;
             }
         }
         return isAlive;
@@ -188,9 +188,9 @@ Flynn.Particles = Class.extend({
 
     },
 
-    update: function(paceFactor) {
+    update: function(elapsedTicks) {
         for(var i=0, len=this.particles.length; i<len; i+=1){
-            if(!this.particles[i].update(paceFactor)){
+            if(!this.particles[i].update(elapsedTicks)){
                 // Particle has died.  Remove it
                 this.particles.splice(i, 1);
                 len--;