+++ js/stateText.js
@@ -16,13 +16,13 @@ Game.StateText = Flynn.State.extend({
         this.rotate_step = -(Math.PI/3)/60;
     },
 
-    handleInputs: function(input, paceFactor) {
+    handleInputs: function(input, elapsedTicks) {
         Game.handleInputs_common(input);
     },
 
-    update: function(paceFactor) {
-        this.gameClock += paceFactor;
-        this.rotate_angle += this.rotate_step * paceFactor;
+    update: function(elapsedTicks) {
+        this.gameClock += elapsedTicks;
+        this.rotate_angle += this.rotate_step * elapsedTicks;
     },
 
     render: function(ctx){